fixed bugs
updated context

+++ server/controllers/userControllers.js
@@ -83,8 +83,13 @@ const getAllUsers = asyncHandler(async (req, res) => {
 const deleteUser = asyncHandler(async (req, res) => {
   const _id = req.params.userId;
   try {
-    const user = await User.findByIdAndDelete(_id);
     if (!user) return res.sendStatus(404);
+
+    if (user.isAdmin)
+      return res.send(user.name + " is admin, you cannot delete an admin");
+
+    const user = await User.findByIdAndDelete(_id);
+
     return res.send(user.name + " with id: " + user._id + " has been deleted!");
   } catch (error) {
     res.status(400);
